backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pytest
pytest-django
psycopg2-binary
tzdata